pyproject.toml
[project]
name = "ct_image_to_pic_list"
version = "0.0.5"
description = "convert a 3d CT/MRI image (numpy) into pillow image list."
authors = [
    {name = "GGN_2015",email = "neko@jlulug.org"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy", "pillow"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
